+++ src/main/java/com/renomad/minum/database/AbstractDb.java
@@ -427,5 +427,4 @@ public abstract class AbstractDb<T extends DbData<?>> {
                     .formatted(emptyInstance.getClass().getSimpleName(), indexName, indexKey));
         }
     }
-
 }