+++ src/test/java/com/renomad/minum/database/DbEngine2Tests.java
@@ -3,6 +3,7 @@ package com.renomad.minum.database;
 import com.renomad.minum.logging.Logger;
 import com.renomad.minum.logging.LoggingLevel;
 import com.renomad.minum.logging.TestLogger;
+import com.renomad.minum.logging.TestLoggerException;
 import com.renomad.minum.state.Constants;
 import com.renomad.minum.state.Context;
 import com.renomad.minum.testing.StopwatchUtils;
@@ -1113,6 +1114,27 @@ public class DbEngine2Tests {
     }
 
     /**
+     * If the data is already loaded, then the loadData method
+     * will not need to be run.
+     * The question is, how do we know whether it was run or not?
+     * One way to tell: if there is no data to load, there will be a log
+     * statement mentioning "adding nothing to the data". If we don't see
+     * that log statement, it means we skipped loadDataFromDisk() successfully.
+     */
+    @Test
+    public void test_LoadData_NoNeed() {
+        Path dbPathForTest = foosDirectory.resolve("test_LoadData_NoNeed");
+        fileUtils.deleteDirectoryRecursivelyIfExists(dbPathForTest);
+        final var db = new DbEngine2<>(dbPathForTest, context, DbTests.Foo.INSTANCE);
+        db.hasLoadedData = true;
+
+        db.loadData();
+
+        assertThrows(TestLoggerException.class,
+                () -> logger.doesMessageExist("Loading data from disk for dbEngine2"));
+    }
+
+    /**
      * If the file to be analyzed as a consolidated file doesn't match
      * expectations, an exception gets thrown.
      */
@@ -1602,10 +1624,7 @@ public class DbEngine2Tests {
         var executorService = Executors.newVirtualThreadPerTaskExecutor();
         var logger = new Logger(constants, executorService, "db_perf_testing");
 
-        var context = new Context(executorService, constants);
-        context.setLogger(logger);
-
-        return context;
+        return new Context(executorService, constants, logger);
     }
 
     /**